+++ CHANGELOG.md
@@ -5,6 +5,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- Add `Diagnostics.stretching_rate`, which allows to compute the stretching
+  rate of vortex filaments based on their instantaneous velocity.
+
 ## [0.22.1] - 2024-06-20
 
 ### Added